Use library for markdown parser
