+++ apps/web/controllers/tasks/create.rb
@@ -38,7 +38,7 @@ module Web::Controllers::Tasks
 
     def task_params
       hash = params[:task]
-      hash[:body] = Markdown.parse(hash[:md_body])
+      hash[:body] = OSSBoard::Markdown.new.parse(hash[:md_body])
       hash[:status] = Task::VALID_STATUSES[:in_progress]
       hash[:approved] = nil
       hash
